Use desktop OpenGL header files on RPI
mpv uses the same OpenGL code for GLES and desktop GL. For this reason,
a set of header files providing a superset of GLES and desktop GL are
required, which, as far as mpv is concerned, are the desktop GL headers.
If desktop GL headers aren't found, all GL code is disabled.

Since the native RPI video output uses GLES for subtitle rendering as of
recently, this is now needed.

OpenELEC so far provides only GLES headers. The easy solution, adding
Mesa to the system, doesn't quite work (colliding header files due to
non-upstream RPI header location), and would needlessly take up more
space on the SD card. So fix it by supplying the Mesa headers to mpv
directly during compilation.

The GL Mesa headers added with this commit were taken from the Mesa git
repository.