Fix build failures on symbol dumping in some cases
This find command picked up "ld.so.conf", which caused symbol dumping to
fail, and with it the entire build process. (It's unknown to me why
others don't have the same problem. Maybe stray files.)

Fix this by strictly only picking up .so files. Most .so files are
actually symlinks to the "proper" versioned .so file, so make it find
symlinks as well by adding "-follow".